- Change the look of the transactions. - Change text of "from" and "to" when it's the user account. Instead of bitshares id, the name of the account will appear.

+++ app/src/main/java/cy/agorise/crystalwallet/viewmodels/CryptoNetAccountViewModel.java
@@ -24,7 +24,7 @@ public class CryptoNetAccountViewModel extends AndroidViewModel {
         this.db = CrystalDatabase.getAppDatabase(application.getApplicationContext());
     }
 
-    public void loadCryptoNetAccount(int accountId){
+    public void loadCryptoNetAccount(long accountId){
         this.cryptoNetAccount = this.db.cryptoNetAccountDao().getByIdLiveData(accountId);
     }
 
@@ -36,5 +36,4 @@ public class CryptoNetAccountViewModel extends AndroidViewModel {
     public LiveData<CryptoNetAccount> getCryptoNetAccount(){
         return this.cryptoNetAccount;
     }
-
 }
